Rebase on the Soteria implementation of JWT; only provide MP specifics
