--- src/main/java/org/omnifaces/jwt/cdi/InjectionPointGenerator.java
@@ -1,70 +1,0 @@
-/*
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
- *
- * Copyright (c) 2017-2021 Payara Foundation and/or its affiliates. All rights reserved.
- *
- * The contents of this file are subject to the terms of either the GNU
- * General Public License Version 2 only ("GPL") or the Common Development
- * and Distribution License("CDDL") (collectively, the "License").  You
- * may not use this file except in compliance with the License.  You can
- * obtain a copy of the License at
- * https://github.com/payara/Payara/blob/master/LICENSE.txt
- * See the License for the specific
- * language governing permissions and limitations under the License.
- *
- * When distributing the software, include this License Header Notice in each
- * file and include the License file at glassfish/legal/LICENSE.txt.
- *
- * GPL Classpath Exception:
- * The Payara Foundation designates this particular file as subject to the "Classpath"
- * exception as provided by the Payara Foundation in the GPL Version 2 section of the License
- * file that accompanied this code.
- *
- * Modifications:
- * If applicable, add the following below the License Header, with the fields
- * enclosed by brackets [] replaced by your own identifying information:
- * "Portions Copyright [year] [name of copyright owner]"
- *
- * Contributor(s):
- * If you wish your version of this file to be governed by only the CDDL or
- * only the GPL Version 2, indicate your decision by adding "[Contributor]
- * elects to include this software in this distribution under the [CDDL or GPL
- * Version 2] license."  If you don't indicate a single choice of license, a
- * recipient has the option to distribute your version of this file under
- * either the CDDL, the GPL Version 2 or to extend the choice of license to
- * its licensees as provided above.  However, if you add GPL Version 2 code
- * and therefore, elected the GPL Version 2 license, then the option applies
- * only if the new code is made subject to such option by the copyright
- * holder.
- */
-// Portions Copyright 2019, 2022 OmniFaces
-package org.omnifaces.jwt.cdi;
-
-import jakarta.enterprise.context.Dependent;
-import jakarta.enterprise.inject.spi.InjectionPoint;
-import jakarta.inject.Inject;
-
-/**
- * Dummy class used to take the injection point for "InjectionPoint" of, for usage in the
- * implementation of
- * <code>BeanManager#getCurrentInjectionPoint(javax.enterprise.context.spi.CreationalContext)</code>.
- * <p>
- * The actual injectionPoint being injected is not used.
- *
- * @author Arjan Tijms
- */
-@Dependent
-public class InjectionPointGenerator {
-
-    // TODO: this is a workaround originally for older OWB versions, but while OWB is fixed, newer Weld
-    // versions are now broken. It seems this needs to be fixed in CDI 2.1, or perhaps CDI 3.
-    // See https://issues.jboss.org/browse/CDI-610
-
-    @Inject
-    private InjectionPoint injectionPoint;
-
-    public InjectionPoint getInjectionPoint() {
-        return injectionPoint;
-    }
-
-}
