+++ src/main/java/org/omnifaces/jwt/mp/jwt/ClaimValueImpl.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) [2021] Payara Foundation and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021 Payara Foundation and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -37,26 +37,36 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package org.omnifaces.jwt.eesecurity;
+package org.omnifaces.jwt.mp.jwt;
+
+import java.util.function.Function;
+import org.eclipse.microprofile.jwt.ClaimValue;
 
 /**
- * Exception during processing of JWT.
+ * A default implementation of {@link ClaimValue}
+ * 
+ * @author Arjan Tijms
+ *
+ * @param <T> the expected type of the claim
  */
-public class JWTProcessingException extends Exception {
-
-    public JWTProcessingException() {
+public class ClaimValueImpl<T> implements ClaimValue<T> {
+    
+    private final String name;
+    private final Function<String, T> valueFunction;
+    
+    public ClaimValueImpl(String name, Function<String, T> valueFunction) {
+        this.name = name;
+        this.valueFunction = valueFunction;
     }
 
-    public JWTProcessingException(String message) {
-        super(message);
+    @Override
+    public String getName() {
+        return name;
     }
 
-    public JWTProcessingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public JWTProcessingException(Throwable cause) {
-        super(cause);
+    @Override
+    public T getValue() {
+        return valueFunction.apply(name);
     }
 
 }
