+++ src/main/java/org/omnifaces/jwt/mp/jwt/JWTInjectableType.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package org.omnifaces.jwt.jwt;
+package org.omnifaces.jwt.mp.jwt;
 
 import static java.util.Collections.singleton;
 import static java.util.stream.Collectors.toSet;
